[refactor][qt] move SendCoinsRecipient to its own file
