+++ src/Makefile.qttest.include
@@ -7,13 +7,13 @@ TESTS += qt/test/test_bitcoin-qt
 
 TEST_QT_MOC_CPP = \
   qt/test/moc_compattests.cpp \
-  qt/test/moc_rpcnestedtests.cpp \
-  qt/test/moc_uritests.cpp
+  qt/test/moc_rpcnestedtests.cpp
 
 if ENABLE_WALLET
 TEST_QT_MOC_CPP += \
   qt/test/moc_paymentservertests.cpp \
-  qt/test/moc_wallettests.cpp
+  qt/test/moc_wallettests.cpp \
+  qt/test/moc_uritests.cpp
 endif
 
 TEST_QT_H = \
@@ -37,12 +37,12 @@ qt_test_test_bitcoin_qt_SOURCES = \
   qt/test/compattests.cpp \
   qt/test/rpcnestedtests.cpp \
   qt/test/test_main.cpp \
-  qt/test/uritests.cpp \
   $(TEST_QT_H) \
   $(TEST_BITCOIN_CPP) \
   $(TEST_BITCOIN_H)
 if ENABLE_WALLET
 qt_test_test_bitcoin_qt_SOURCES += \
+  qt/test/uritests.cpp \
   qt/test/paymentservertests.cpp \
   qt/test/wallettests.cpp \
   wallet/test/wallet_test_fixture.cpp
